pyproject.toml
[tool.poetry]
name = "hsutools"
version = "1.1.0"
description = "Utilities for directory readme generation, file management, renaming, and docx to pdf conversion."
authors = ["hsutools maintainers"]
readme = "README.md"
license = "MIT"
packages = [{ include = "hsutools", from = "src" }]

[tool.poetry.dependencies]
python = ">=3.10,<3.14"
typer = {version = "^0.12.5", extras = ["all"]}
click = ">=8.1.7,<8.2"
docx2pdf = "^0.1.8"
tabulate = "^0.9.0"
pillow = ">=10.0,<12.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"
pyinstaller = "^6.11.0"

[tool.poetry.scripts]
hsu = "hsutools.cli:app"

[build-system]
requires = ["poetry-core>=1.9.0"]
build-backend = "poetry.core.masonry.api"
